# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Productivity Analytics Platform API
  version: 1.0.0
  description: "\n## Overview\n\nA comprehensive RESTful API for intelligent task
    management, habit tracking, \nand productivity analytics. Built with Django REST
    Framework.\n\n## Features\n\n- **Task Management**: Full CRUD operations with
    categories, tags, and priorities\n- **Habit Tracking**: Daily/weekly habits with
    streak tracking\n- **Pomodoro Timer**: Focus session management and tracking\n-
    **Productivity Analytics**: Daily snapshots, trends, and insights\n- **Weather
    Integration**: Correlate productivity with weather conditions\n\n## Authentication\n\nThis
    API uses JWT (JSON Web Token) authentication. Include the token in the Authorization
    header:\n\n```\nAuthorization: Bearer <your_access_token>\n```\n\n## Rate Limiting\n\n-
    Anonymous users: 100 requests/hour\n- Authenticated users: 1000 requests/hour\n\n##
    MCP Compatibility\n\nThis API is designed to be Model Context Protocol (MCP) compatible
    for AI agent integration.\n    "
  contact:
    name: API Support
    email: support@productivity-api.example.com
  license:
    name: MIT License
paths:
...
  /api/v1/auth/register/:
    post:
      operationId: auth_register_create
      description: "\nRegister a new user account.\n\nReturns JWT tokens (access and
        refresh) upon successful registration.\n\n**Password Requirements:**\n- Minimum
        8 characters\n- Cannot be entirely numeric\n- Cannot be too similar to other
        personal information\n- Cannot be a commonly used password\n        "
      summary: Register a new user
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                  tokens:
                    type: object
                    properties:
                      access:
                        type: string
                      refresh:
                        type: string
                  message:
                    type: string
          description: ''
  /api/v1/auth/token/:
    post:
      operationId: auth_token_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPairRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenObtainPairRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenObtainPairRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ''
...
  /api/v1/auth/token/refresh/:
    post:
      operationId: auth_token_refresh_create
      description: |-
        Takes a refresh type JSON web token and returns an access type JSON web
        token if the refresh token is valid.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
...
  /api/v1/mcp/execute/:
    post:
      operationId: mcp_execute_create
      description: "\nExecute a tool by name with the provided arguments.\n\n**Request
        Format:**\n```json\n{\n    \"tool\": \"tool_name\",\n    \"arguments\": {\n
        \       \"param1\": \"value1\",\n        \"param2\": \"value2\"\n    }\n}\n```\n\n**Response
        Format:**\n```json\n{\n    \"success\": true,\n    \"result\": { ... },\n
        \   \"error\": null\n}\n```\n        "
      summary: Execute an MCP tool
      tags:
      - MCP
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPExecuteRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/MCPExecuteRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MCPExecuteRequestRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPExecuteResponse'
          description: ''
  /api/v1/mcp/tools/:
    get:
      operationId: mcp_tools_retrieve
      description: "\nReturns all available tools/operations that can be executed
        through the MCP interface.\n\nEach tool includes:\n- Name and description\n-
        Input schema (JSON Schema format)\n- Required authentication level\n        "
      summary: List available MCP tools
      tags:
      - MCP
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
          description: ''
...
components:
  schemas:
...
    MCPExecuteRequestRequest:
      type: object
      description: Request body for MCP tool execution.
      properties:
        tool:
          type: string
          minLength: 1
          description: Name of the tool to execute
        arguments:
          type: object
          additionalProperties: {}
          description: Arguments for the tool
      required:
      - tool
    MCPExecuteResponse:
      type: object
      description: Response for MCP tool execution.
      properties:
        success:
          type: boolean
          description: Whether the execution was successful
        result:
          type: object
          additionalProperties: {}
          nullable: true
          description: Result data from the tool execution
        error:
          type: string
          nullable: true
          description: Error message if execution failed
      required:
      - success
...
    TokenObtainPair:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - refresh
    TokenObtainPairRequest:
      type: object
      properties:
        email:
          type: string
          writeOnly: true
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - password
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
    TokenRefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
...
    UserRegistrationRequest:
      type: object
      description: |-
        Serializer for user registration.

        Validates email uniqueness, password strength, and password confirmation.
      properties:
        email:
          type: string
          format: email
          minLength: 1
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 1
        password_confirm:
          type: string
          writeOnly: true
          minLength: 1
        timezone:
          type: string
          minLength: 1
          maxLength: 50
        preferred_work_start:
          type: string
          format: time
        preferred_work_end:
          type: string
          format: time
        daily_task_goal:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
      required:
      - email
      - password
      - password_confirm
      - username
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...